fix : PaymentUser의 id 타입 수정

+++ src/main/java/com/system/payment/user/domain/jaebin/PaymentUser.java
@@ -5,8 +5,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Table(name = "payment_user")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -15,7 +13,7 @@ public class PaymentUser extends BaseTimeEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false, unique = true, length = 200)
     private String email;
